Flatten Test directory structure
