--- src/Tests/CoreTests/main.cpp
@@ -1,20 +1,0 @@
-#include <Tests/CoreTests/Tests.hpp>
-
-#include <Tests/CoreTests/Algebra/AlgebraTests.hpp>
-#include <Tests/CoreTests/Animation/AnimationTest.hpp>
-#include <Tests/CoreTests/Containers/ContainersTest.hpp>
-#include <Tests/CoreTests/Containers/IndexMapTest.hpp>
-#include <Tests/CoreTests/Distance/DistanceTests.hpp>
-#include <Tests/CoreTests/Geometry/GeometryTests.hpp>
-#include <Tests/CoreTests/RayCasts/RayCastTest.hpp>
-#include <Tests/CoreTests/String/StringTest.hpp>
-#include <Tests/CoreTests/TopologicalMesh/ConvertTest.hpp>
-
-int main() {
-    if ( !RaTests::TestManager::getInstance() )
-    {
-        RaTests::TestManager::createInstance();
-    }
-    RaTests::TestManager::getInstance()->m_options.m_breakOnFailure = true;
-    return RaTests::TestManager::getInstance()->run();
-}
