+++ src/Tests/CoreTests/src/RayCastTest.hpp
@@ -2,7 +2,7 @@
 #define RADIUM_RAYCAST_TESTS_HPP_
 
 #include <Core/Math/RayCast.hpp>
-#include <Tests/CoreTests/Tests.hpp>
+#include <Tests/Tests.hpp>
 
 namespace RaTests {
 
@@ -42,7 +42,8 @@ class RayCastAabbTests : public Test {
                                 RA_UNIT_TEST(
                                     Ra::Core::Math::areApproxEqual( r.pointAt( t )[i], sig ),
                                     "Wrong hit point" );
-                            } else
+                            }
+                            else
                             { RA_UNIT_TEST( !result, "The ray should have missed" ); }
                         }
