+++ src/Tests/Manager.hpp
